script: Do not allow a delimiter between function name and block start.
Currently the following is valid syntax, but should be a syntax error:

grub> function f; { echo HERE; }
grub> f
HERE

This fix is not backward compatible, but current syntax is not documented
either and has no functional value. So any scripts with this unintended
syntax are technically syntactically incorrect and should not be relying
on this behavior.