parser: Enable script parser's debug output when setting "parser" debug conditional.
This can be useful when tracking down parsing and scripting issues. There
are two code blocks depending on GRUB_UTIL. If GRUB_UTIL is set, then we
are in a normal posix environment and have stdio.h.  If not, we are
compiling for the grub environment, in which case we ignore the FILE
argument and set up a fake stdio env.