luks: When copying UUID use sizeof (header.uuid), since its guaranteed to be smaller.
The COMPILE_TIME_ASSERT assures that sizeof (header.uuid) is smaller than
sizeof (newdev->uuid), so we do not need to copy more bytes than is possible
for the UUID. Because header.uuid might not be NULL terminated, we might not
end up with a NULL terminated string in newdev->uuid. However since newdev
was just zalloc'd and sizeof (newdev->uuid) is larger than sizeof
(header.uuid), there is guaranteed to be a NULL in newdev->uuid.