cryptodisk: Add ability to specify large IV encryption option, and return unsupported.
Cryptsetup has the ability to setup encrypted containers with the large IV
option, but only as a plain container (ie not available with a LUKS header).
This is where a plain IV increments once for every logical encrypted block.
Normally, the IV increments once for every DM-crypt native block (512-bytes).