misc: Remove grub_strcasecmp and replace with a macro to grub_strncasecmp.
This eliminates essentially repeated code in misc.h because grub_strcasecmp
is really a special case of grub_strncasecmp.  The only reason do have an
extra function would be for performance reasons, but not much is being saved
here, a few bytes of stack, a few memory read/writes and a few branches.