misc: Change grub_strword to return a string like grub_strstr.
This makes grub_strword more versatile. All code as of this change using
grub_strword uses it in a conditional or boolean operator context where its
return value would be sliently converted to a 1 or 0, so this does not
affect current usage. I have opted to change a few lines in loader/xnu.c to
enhance clarity.